Move argument definitions into header file and define with macros.
- Arguments are defined in the new file "arg_definitions.h" and used by
  including this file with appropriate macro definitions.
- Add a "-b|--bash-completions" option which prints bash completions to
  stdout. Use `eval $(astroterm -b)` to enable them in bash.

Explanation: I could have copy-pasted the arguments into the bash
completions code, but this would have made it more brittle to add/change
arguments in the future. This approach means you have a single place
where arguments are defined, and they can be re-used with macros. This
approach of iterating over cities also means that if the list of cities
changes in the future, the completions will remain up-to-date.