Update nix installation instructions
It is generally discouraged to use nix-env to install packages:
"""
Warning: Using nix-env permanently modifies a local profile of installed packages. This must be updated and maintained by the user in the same way as with a traditional package manager, foregoing many of the benefits that make Nix uniquely powerful. Using nix-shell or a NixOS configuration is recommended instead. 
"""
See for example 
https://search.nixos.org/packages?channel=25.05&show=astroterm&from=0&size=50&sort=relevance&type=packages&query=astroterm